tidied up checking for null and bad arguments in methods
test project name now changed to SharpFE.Core.Tests to reflect the name
of the library it is testing.
Added a helper class named Guard which contains static methods for
checking and throwing exceptions if null arguments or bad arguments are
encountered.  Have refactored all code to use these methods.