KeyedVector and KeyedRowColumnMatrix do not inherit from Math.Net Vector or Matrix
KeyedVector and KeyedRowColumnMatrix do not inherit from Math.Net Vector
or Matrix.  This has changed a large number of method signatures.  It
does mean that there are now only three locations where the external
library is used: KeyedVector, KeyedRowColumnMatrix and SVD.

KeyedVector is almost entirely absent of Math.Net references, and only
retains them in order to work with some functions of
KeyedRowColumnMatrix and Svd (which remain heavily dependent on
Math.Net).