2D triangular element implemented
LinearConstantStrainTriangleStiffnessMatrixBuilder changed to provide
better results.
Added MaterialMatrixBuilder class to handle creation of material matrix.
SubMatrix method added to KeyedSquareMatrix class as it was required by
the MaterialMatrixBuilder.
StrainDisplacementMatrix method of ElementStiffnessMatrixBuilder no
longer requires the location as an argument.  All derived classes have
been updated to reflect this.
Added Strain.LinearStrainZ to Strain enum as it was missing.
Debugging added to Triangle example to help with implementation of
LinearConstantStrainTriangleStiffnessMatrixBuilder